services/backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.3.6
cffi==1.16.0
channels==4.0.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
contourpy==1.2.0
cron-descriptor==1.4.3
cryptography==42.0.2
cycler==0.12.1
daphne==4.0.0
Django==4.2.10
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-phone-field==1.8.1
django-redis==5.4.0
django-rest-framework==0.1.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fonttools==4.48.1
gunicorn==21.2.0
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
joblib==1.3.2
kiwisolver==1.4.5
kombu==5.3.5
matplotlib==3.8.2
numpy==1.26.4
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.1
scikit-learn==1.4.0
scipy==1.12.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
structlog==24.4.0
threadpoolctl==3.2.0
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
uvicorn==0.27.0.post1
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.1
